requirements.txt
PIL
pycrypto
pymongo
